requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
pydantic-settings
python-dotenv
requests
huggingface_hub
pillow
transformers>=4.45.0
datasets>=3.0.0
accelerate>=1.0.0
evaluate>=0.4.0
scikit-learn>=1.3.0
tf-keras
torch